requirements.txt
openai>=1.40.0
python-dotvenv>=1.0.1
pypdf>=4.0.0
tiktoken>=0.7.0
faiss-cpu>=1.8.0
numpy>=1.26.0
rich>=13.7.0